Database class
Added SQLite Database class that reads and writes into the database.
Settings has been changed in order to add strings. Ideally this would be
subreddits which could be used. No way to grab individual subreddits yet
or delete database. App has to be uninstalled and reinstalled in order
to delete database.